Program.cs:
Added in the choices for the user to chose which application that they want to go into.
Added a promt that would appear to encourage the user to make a choice "Please chose an application..."

DistanceConverter:

Added in the summary for the application.

Added in the Distance converter class which contains the distance variables and the conversion values for those variables.